+++ String Algorithms/String_Hashing.cpp
@@ -2,10 +2,10 @@ struct Hashing{
     string s;
     int n;
     int primes;
-    vector<ll> hashPrimes = {1000000009, 100000007};
-    const ll base = 31;
-    vector<vector<ll>> hashValues;
-    vector<vector<ll>> powersOfBase;
+    vector<long long> hashPrimes = {1000000009, 100000007};
+    const long long base = 31;
+    vector<vector<long long>> hashValues;
+    vector<vector<long long>> powersOfBase;
     Hashing(string a){
         primes = sz(hashPrimes);
         hashValues.resize(primes);
@@ -40,17 +40,17 @@ struct Hashing{
                 if(sz(hashValues[i]) == 0){
                     hashValues[i].pb((s[0] - 'a' + 1LL) % hashPrimes[i]);
                 }else{
-                    ll extraHash = hashValues[i].back() + ((s.back() - 'a' + 1LL) * powersOfBase[i][sz(s) - 1]) % hashPrimes[i];
+                    long long extraHash = hashValues[i].back() + ((s.back() - 'a' + 1LL) * powersOfBase[i][sz(s) - 1]) % hashPrimes[i];
                     hashValues[i].pb((extraHash + hashPrimes[i]) % hashPrimes[i]);
                 }
             }
         }
     }
-    vector<ll> substringHash(int l, int r){ // extra O(log) factor
-        vector<ll> hash(primes);
+    vector<long long> substringHash(int l, int r){ // extra O(log) factor
+        vector<long long> hash(primes);
         for(int i = 0; i < primes; i++){
-            ll val1 = hashValues[i][r];
-            ll val2 = l > 0 ? hashValues[i][l - 1] : 0LL;
+            long long val1 = hashValues[i][r];
+            long long val2 = l > 0 ? hashValues[i][l - 1] : 0LL;
             hash[i] = mod_mul(mod_sub(val1, val2, hashPrimes[i]), mminvprime(powersOfBase[i][l], hashPrimes[i]), hashPrimes[i]);
         }
         return hash;
@@ -61,8 +61,8 @@ struct Hashing{
             swap(r1, r2);
         }
         for(int i = 0; i < primes; i++){
-            ll val1 = mod_sub(hashValues[i][r1], (l1 > 0 ? hashValues[i][l1 - 1] : 0LL), hashPrimes[i]);
-            ll val2 = mod_sub(hashValues[i][r2], (l2 > 0 ? hashValues[i][l2 - 1] : 0LL), hashPrimes[i]);
+            long long val1 = mod_sub(hashValues[i][r1], (l1 > 0 ? hashValues[i][l1 - 1] : 0LL), hashPrimes[i]);
+            long long val2 = mod_sub(hashValues[i][r2], (l2 > 0 ? hashValues[i][l2 - 1] : 0LL), hashPrimes[i]);
             if(mod_mul(val1, powersOfBase[i][l2 - l1], hashPrimes[i]) != val2)
                 return false;
         }
